feat: `createStyleguideConfig` now supports custom sections
BREAKING CHANGE: `crs.config.js` has been removed in favor of passing options directly to `createStyleguideConfig`.
`createStyleguideConfig` now takes two params, the first is a config object that will be shallowly merged,
the second is an options object that takes the old `styleguides` option for the removed `crs.config.js`.
A few other options have been added for customizing how sections are displayed, and this also works with
the `DEBUG=true` environmental variable.